requirements.txt
django
django-debug-toolbar
djangorestframework
markdown
django-filter
